requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
datasets>=2.12.0
accelerate>=0.20.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
wandb>=0.15.0
pyyaml>=6.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0